% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mantelhaenszel.R
\name{add_mantelhaenszel}
\alias{add_mantelhaenszel}
\title{Creates a raschtreeMH object based on the original Raschtree with additional Mantel-Haenszel effect size information}
\usage{
add_mantelhaenszel(object, purification)
}
\arguments{
\item{object}{An object of type Raschtree}

\item{purification}{A character indicating the type of purification ("none", "2step", "iterative")}
}
\value{
An object of type raschtreeMH with 'info' extended by a list named mantelhaenszel containing entries Mantel-Haenszel effect size, classification, purification type, and purificationCounter
}
\description{
Creates a raschtreeMH object based on the original Raschtree with additional Mantel-Haenszel effect size information
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim,
                stopfun = stopfun_mantelhaenszel(purification = "iterative"))
RT_MH <- add_mantelhaenszel(RT, purification = "iterative")
RT_MH$info$mantelhaenszel
plot(RT_MH, color_by_node = 1)
}
}
