% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mantelHaenszel.R
\name{calculate_mantelHaenszel}
\alias{calculate_mantelHaenszel}
\title{Calculates the Mantel-Haenszel statistic for Differential Item Functioning (DIF)}
\usage{
calculate_mantelHaenszel(dat, splitGroup, sums, purification, ...)
}
\arguments{
\item{dat}{A dataframe with dichotomous item responses}

\item{splitGroup}{An indicator of the two groups for which DIF analyses should be performed}

\item{sums}{A sum score used as a matching criterion for computing the Mantel Haenszel statistic}

\item{purification}{A character indicating the type of purification that is used on the Mantel Haenszel statistic. Options are "none", "2step", or "iterative"}

\item{...}{Further arguments}
}
\value{
A list with the Mantel-Haenszel effect size measure in the Delta scale, the classification based on the ETS guidelines and information on purification and convergence of purification if iterative purification was done.
}
\description{
Calculates the Mantel-Haenszel statistic for Differential Item Functioning (DIF)
}
