% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_node.R
\name{color_by_node}
\alias{color_by_node}
\title{Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification}
\usage{
color_by_node(node_ID, class_color, class_size, panel_color)
}
\arguments{
\item{node_ID}{An integer indicating the inner node after which the item paramter should be colored}

\item{class_color}{A character vector indicating the colors for items classified as A, B, or C}

\item{class_size}{A character vector indicating the size for items classified as A, B, or C}

\item{panel_color}{A character vector indicating the colors of the background panels for left and right terminal nodes of the inner node by which items are colored}
}
\value{
A function that takes the argument 'purification'node_ID' and 'color' and can be used as a value for the argument 'terminal_panel' in plot.raschtree()
}
\description{
Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification
}
