% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_byNode.R
\name{color_byNode}
\alias{color_byNode}
\title{Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification}
\usage{
color_byNode(object, nodeName, cols = c("green", "orange", "red"))
}
\arguments{
\item{nodeName}{An integer indicating the inner node after which the item paramter should be colored}

\item{cols}{A named character vector indicating the}
}
\value{
A list named after the terminal nodes containing the colors of the item parameter based on ETS Mantel-Haenszel classification
}
\description{
Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification
}
\examples{
data("DIFSim", package = "psychotree")
rt <- raschtree(resp ~ age + gender + motivation, data = DIFSim, verbose = TRUE)
rt_MH <- add_mantelHaenszel(rt, purification = "iterative")
plot(rt_MH, col = color_byNode(rt_MH, 2))

}
