% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_node.R
\name{create_bg_list}
\alias{create_bg_list}
\title{Helper function to create a list with colors for left and right child node of the node by which A/B/C coloring is performed}
\usage{
create_bg_list(object, node_ID, background_cols)
}
\arguments{
\item{object}{An object of class raschtree that has the mantelhaenszel statistic added to it.}

\item{node_ID}{An integer indicating the inner node after which the item paramter should be colored}

\item{background_cols}{A character vector indicating the colors of the background panels for left and right terminal nodes of the inner node by which items are colored}
}
\value{
A list named after the terminal nodes containing the colors of the item parameter based on ETS Mantel-Haenszel classification
}
\description{
Helper function to create a list with colors for left and right child node of the node by which A/B/C coloring is performed
}
