% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_byNode.R
\name{create_bgList}
\alias{create_bgList}
\title{Helper function to create a list with colors for left and right child node of the node by which A/B/C coloring is performed}
\usage{
create_bgList(object, nodeID, backgroundCols)
}
\arguments{
\item{object}{An object of class raschtree that has the mantelHaenszel statistic added to it.}

\item{nodeID}{An integer indicating the inner node after which the item paramter should be colored}

\item{backgroundCols}{A character vector indicating the colors of the background panels for left and right terminal nodes of the inner node by which items are colored}
}
\value{
A list named after the terminal nodes containing the colors of the item parameter based on ETS Mantel-Haenszel classification
}
\description{
Helper function to create a list with colors for left and right child node of the node by which A/B/C coloring is performed
}
