% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ETS_classification.R
\name{get_ETS_classification}
\alias{get_ETS_classification}
\title{Performs classification of the Mantel-Haenszel effect size measured based on ETS guidelines
see Dorans & Holland, 1993, p. 41f or Paek & Holland, 2015, Psychometrika}
\usage{
get_ETS_classification(res_MH, res_MH_sd)
}
\arguments{
\item{res_MH}{A vector or dataframe with the Mantel-Haenszel Odds Ratio effect size measure in the Delta scale (-2.35 * log (Mantel-Haenszel estimate of the common odds ratios)) for each item}

\item{res_MH_sd}{A vector or dataframe with the standard deviation of Mantel-Haenszel Odds Ratio effect size measure in the Delta scale for each item (-2.35 * sqrt (the variance of the lambda_MH statistic))}
}
\value{
A vector or dataframe with the corresponding A/B/C classification for each item
Computes p-values for the statistical hypothesis testing based on log transformation of the Mantel-Haenszel common odds ratio in the Delta-scale for DIF
p values of the hypothesis test are need for the classification of DIF items based on the ETS guidelines
based on Paek & Holland, 2015, Psychometrika
}
\description{
Performs classification of the Mantel-Haenszel effect size measured based on ETS guidelines
see Dorans & Holland, 1993, p. 41f or Paek & Holland, 2015, Psychometrika
}
