% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_split_groups.R
\name{get_inner_nodes}
\alias{get_inner_nodes}
\title{Function that returns the inner nodes of a tree object}
\usage{
get_inner_nodes(node)
}
\arguments{
\item{node}{A node in a Raschtree}
}
\value{
inner nodes of a tree object
}
\description{
Function that returns the inner nodes of a tree object
}
