% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_innerNodes.R
\name{get_innerNodes}
\alias{get_innerNodes}
\title{Function that returns the inner nodes of a tree object}
\usage{
get_innerNodes(node)
}
\description{
Function that returns the inner nodes of a tree object
}
