% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mantelHaenszel.R
\name{get_mantelHaenszel}
\alias{get_mantelHaenszel}
\title{Function that returns the Mantel-Haenszel effect size measure of a Raschtree object}
\usage{
get_mantelHaenszel(object, purification, by = "node")
}
\arguments{
\item{object}{An object of type Raschtree}

\item{purification}{A character indicating the type of purification ("none", "2step", "iterative")}

\item{by}{A character either "node" or "type" defining the structure of the function output}
}
\value{
If by = "node": a list with one entry for each node; If by = "type": a list with entries Mantel-Haenszel effect size, classification, purification type, and purificationCounter
}
\description{
Function that returns the Mantel-Haenszel effect size measure of a Raschtree object
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
MH <- get_mantelHaenszel(RT, purification = "iterative", by = "type")
MH
}

}
