% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MHpVal.R
\name{get_MHpVal}
\alias{get_MHpVal}
\title{Computes p-values for the statistical hypothesis testing based on log transformation of the Mantel-Haenszel common odds ratio in the Delta-scale for DIF
p values of the hypothesis test are need for the classification of DIF items based on the ETS guidelines
based on Paek & Holland, 2015, Psychometrika}
\usage{
get_MHpVal(resMH, resMHsd, tau = 1)
}
\description{
Computes p-values for the statistical hypothesis testing based on log transformation of the Mantel-Haenszel common odds ratio in the Delta-scale for DIF
p values of the hypothesis test are need for the classification of DIF items based on the ETS guidelines
based on Paek & Holland, 2015, Psychometrika
}
