% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splitGroup.R
\name{get_splitGroup}
\alias{get_splitGroup}
\title{Returns the split groups for a node in a tree object}
\usage{
get_splitGroup(node, datFitted)
}
\description{
Returns the split groups for a node in a tree object
}
