% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_split_groups.R
\name{get_terminal_nodes}
\alias{get_terminal_nodes}
\title{Returns the terminal nodes of a tree object}
\usage{
get_terminal_nodes(node)
}
\arguments{
\item{node}{A node in a Raschtree}
}
\value{
terminal nodes of a tree object
}
\description{
Returns the terminal nodes of a tree object
}
