% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_splitGroups.R
\name{get_terminalNodes}
\alias{get_terminalNodes}
\title{Returns the terminal nodes of a tree object}
\usage{
get_terminalNodes(node)
}
\arguments{
\item{node}{A node in a Raschtree}
}
\value{
terminal nodes of a tree object
}
\description{
Returns the terminal nodes of a tree object
}
