% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_split_groups.R
\name{is_terminal_node}
\alias{is_terminal_node}
\title{checks whether a node is a terminal node}
\usage{
is_terminal_node(node)
}
\arguments{
\item{node}{A node in a Raschtree}
}
\description{
checks whether a node is a terminal node
}
