% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_terminalNodes.R
\name{is_terminalNode}
\alias{is_terminalNode}
\title{checks whether a node is a terminal node}
\usage{
is_terminalNode(node)
}
\arguments{
\item{node}{A node in a Raschtree}
}
\description{
checks whether a node is a terminal node
}
