% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparty.R
\name{mob}
\alias{mob}
\title{MOB is an algorithm for model-based recursive partitioning yielding a tree with fitted models associated with each terminal node.
for more details see ?partykit::mob}
\usage{
mob(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  cluster,
  fit,
  control = mob_control(),
  ...
)
}
\arguments{
\item{formula}{symbolic description of the model}

\item{data}{arguments controlling formula processing via model.frame}

\item{subset}{arguments controlling formula processing via model.frame}

\item{na.action}{arguments controlling formula processing via model.frame}

\item{weights}{optional numeric vector of weights. By default these are treated as case weights but the default can be changed in mob_control}

\item{offset}{optional numeric vector with an a priori known component to be included in the model y ~ x1 + ... + xk (i.e., only when x variables are specified)}

\item{cluster}{optional vector (typically numeric or factor) with a cluster ID to be passed on to the fit function and employed for clustered covariances in the parameter stability tests.}

\item{fit}{function. A function for fitting the model within each node. For details see below.}

\item{control}{A list with control parameters as returned by mob_control}

\item{...}{Additional arguments passed to the fit function}
}
\description{
MOB is an algorithm for model-based recursive partitioning yielding a tree with fitted models associated with each terminal node.
for more details see ?partykit::mob
}
