% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschtree.R
\name{raschtree}
\alias{raschtree}
\title{Recursive partitioning (also known as trees) based on Rasch models.
for more details see psychotree::raschtree}
\usage{
raschtree(
  formula,
  data,
  na.action,
  reltol = 1e-10,
  deriv = c("sum", "diff", "numeric"),
  maxit = 100L,
  ...
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. This should be of type y ~ x1 + x2 where y should be a binary 0/1 item response matrix and x1 and x2 are used as partitioning variables.}

\item{data}{a data frame containing the variables in the model.}

\item{na.action}{a function which indicates what should happen when the data contain missing values (NAs).}

\item{reltol}{arguments passed via raschmodel to optim.}

\item{deriv}{character. Which type of derivatives should be used for computing gradient and Hessian matrix? Analytical with sum algorithm ("sum"), analytical with difference algorithm ("diff", faster but numerically unstable), or numerical. Passed to raschmodel.}

\item{maxit}{arguments passed via raschmodel to optim.}

\item{...}{arguments passed to the underlying functions, i.e., to mob_control for raschtree, and to the underlying predict and plot methods, respectively.}
}
\description{
Recursive partitioning (also known as trees) based on Rasch models.
for more details see psychotree::raschtree
}
