% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschtree_mantelhaenszel.R
\name{raschtree_mantelhaenszel}
\alias{raschtree_mantelhaenszel}
\title{Fits raschtree based on the psychotree package with additional options for stopping based on the Mantel-Haenszel effect size measure}
\usage{
raschtree_mantelhaenszel(
  formula,
  data,
  stopping = TRUE,
  stopkrit = "A",
  purification,
  ...
)
}
\arguments{
\item{stopping}{A logical indicator whether stopping based on the Mantel-Haenszel effect size measure should be performed. Default is TRUE}

\item{stopkrit}{A character indicating the stopping criterion. Options are "A", "B", "C" or combinations of these. Stopping is done when all items in follow-up nodes fall into the stopping criterion.}

\item{purification}{A character indicating the type of purification that should be done. Options are "none", "2step", or "iterative".}
}
\value{
A list with two entries: raschtree object and information on the Mantel-Haenszel statistic
}
\description{
Fits raschtree based on the psychotree package with additional options for stopping based on the Mantel-Haenszel effect size measure
}
