% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschTreeEffectSize-package.R
\docType{package}
\name{raschTreeEffectSize-package}
\alias{raschTreeEffectSize-package}
\alias{raschTreeEffectSize}
\title{raschTreeEffectSize: Helper functions to use the Mantel-Haenszel effect size measure as a stopping criterion in Raschtrees}
\description{
Helper functions to use the Mantel-Haenszel effect size measure as a stopping criterion in Raschtrees
}
\author{
Mirka Henninger
}
