% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschtreeMH-package.R
\docType{package}
\name{raschtreeMH-package}
\alias{raschtreeMH-package}
\alias{raschtreeMH}
\title{raschtreeMH: Helper functions to use the Mantel-Haenszel effect size measure as a stopping criterion in Rasch trees}
\description{
Helper functions to use the Mantel-Haenszel effect size measure as a stopping criterion in Rasch trees
}
\author{
Mirka Henninger
}
