% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ETSMH.R
\name{show_ETSMH}
\alias{show_ETSMH}
\title{Helper function to display the ETS Mantel-Haenszel classification summary in each inner node of the Raschtree}
\usage{
show_ETSMH(purification)
}
\arguments{
\item{purification}{A character indicating the type of purification ("none", "2step", "iterative"}
}
\value{
A function that takes the argument purification and can be used as a value for the argument 'inner_panel' in plot.raschtree()
}
\description{
Helper function to display the ETS Mantel-Haenszel classification summary in each inner node of the Raschtree
}
\examples{
data("DIFSim", package = "psychotree")
rt <- raschtree(resp ~ age + gender + motivation, data = DIFSim, verbose = TRUE)
plot(rt, inner_panel = show_ETSMH(purification = "iterative"))

}
