% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mantelhaenszel.R
\name{summary_mantelhaenszel}
\alias{summary_mantelhaenszel}
\title{Improves the output of the Mantel Haenszel statistic for the raschtree}
\usage{
summary_mantelhaenszel(x)
}
\arguments{
\item{x}{An object (list) returned from the get_mantelHaenszel.R function}
}
\value{
A list with classification, Mantel-Haenszel odds ratio, purification, and purification counter for each split in the Raschtree
}
\description{
Improves the output of the Mantel Haenszel statistic for the raschtree
}
