% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mantelHaenszel.R
\name{add_mantelHaenszel}
\alias{add_mantelHaenszel}
\title{Function that saves Mantel-Haenszel effect size measures of a Raschtree into the Raschtree object}
\usage{
add_mantelHaenszel(object, purification)
}
\arguments{
\item{object}{An object of type Raschtree}

\item{purification}{A character indicating the type of purification ("none", "2step", "iterative")}
}
\value{
The Raschtree object with 'info' extended by a list named mantelHaenszel containing entries Mantel-Haenszel effect size, classification, purification type, and purificationCounter
}
\description{
Function that saves Mantel-Haenszel effect size measures of a Raschtree into the Raschtree object
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
RT_MH <- add_mantelHaenszel(RT, purification = "iterative")
RT_MH$info$mantelHaenszel
}
}
