% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_node.R
\name{anchor_pars}
\alias{anchor_pars}
\title{Helper function for anchoring items in the Rasch tree profile plot}
\usage{
anchor_pars(node_ID)
}
\arguments{
\item{node_ID}{An integer indicating the inner node after which the item paramter should be colored}
}
\value{
function cf that will be used to extract anchored item parameter from the mob object
}
\description{
Helper function for anchoring items in the Rasch tree profile plot
}
