% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_byNode.R
\name{color_byNode}
\alias{color_byNode}
\title{Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification}
\usage{
color_byNode(nodeID, classColor, classSize, panelColor)
}
\arguments{
\item{nodeID}{An integer indicating the inner node after which the item paramter should be colored}

\item{classColor}{A character vector indicating the colors for items classified as A, B, or C}

\item{classSize}{A character vector indicating the size for items classified as A, B, or C}

\item{panelColor}{A character vector indicating the colors of the background panels for left and right terminal nodes of the inner node by which items are colored}
}
\value{
A function that takes the argument 'purification'nodeID' and 'color' and can be used as a value for the argument 'terminal_panel' in plot.raschtree()
}
\description{
Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification
}
