% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_byNode.R
\name{color_byNode}
\alias{color_byNode}
\title{Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification}
\usage{
color_byNode(nodeID, color = c("green", "orange", "red"))
}
\arguments{
\item{nodeID}{An integer indicating the inner node after which the item paramter should be colored}

\item{color}{A character vector indicating the colors for items classified as A, B, or C}
}
\value{
A function that takes the argument 'purification'nodeID' and 'color' and can be used as a value for the argument 'terminal_panel' in plot.raschtree()
}
\description{
Helper function to to be used for the 'col' argument in the plot method for Raschtrees
Helps to color item parameter in profileplots by inner nodes and ETS Mantel-Haenszel classification
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim, verbose = TRUE)
RT_MH <- add_mantelHaenszel(RT, purification = "iterative")
plot(RT_MH, terminal_panel = color_byNode(nodeID = 1, color = c("blue", "orange", "darkgreen")))
}

}
