% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_byNode.R
\name{create_colorList}
\alias{create_colorList}
\title{Helper function to create a list with colors for each terminal node}
\usage{
create_colorList(object, nodeID, cols)
}
\arguments{
\item{object}{An object of class raschtree that has the mantelHaenszel statistic added to it.}

\item{nodeID}{An integer indicating the inner node after which the item paramter should be colored}

\item{cols}{A named character vector indicating the}
}
\value{
A list named after the terminal nodes containing the colors of the item parameter based on ETS Mantel-Haenszel classification
}
\description{
Helper function to create a list with colors for each terminal node
}
