% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mantelHaenszel.R
\name{get_mantelHaenszel}
\alias{get_mantelHaenszel}
\title{Function that returns the Mantel-Haenszel effect size measure of a Raschtree object}
\usage{
get_mantelHaenszel(object, purification, ...)
}
\arguments{
\item{object}{An object of type Raschtree}

\item{purification}{A character indicating the type of purification ("none", "2step", "iterative")}
}
\value{
A list with Mantel-Haenszel effect size, classification, purification type, and convergence if purification was iterative
}
\description{
Function that returns the Mantel-Haenszel effect size measure of a Raschtree object
}
