% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mantelhaenszel.R
\name{plot.raschtreeMH}
\alias{plot.raschtreeMH}
\title{Plots the Mantel-Haenszel Raschtree}
\usage{
\method{plot}{raschtreeMH}(
  x,
  show_classification = TRUE,
  color_by_node = NULL,
  ABC_colors = c("#99e1e3", "#ba6100", "#6c0200"),
  ABC_size = c(0.4, 0.75, 0.9),
  node_background = c("#eee3af", "#aeaeae"),
  ...
)
}
\arguments{
\item{x}{An object of type raschtreeMH with 'info' extended by a list named mantelhaenszel containing entries Mantel-Haenszel effect size, classification, purification type, and purificationCounter}

\item{show_classification}{A logical with default is TRUE: Should the ETS-Mantel-Haenszel classification be shown in each inner node?}

\item{color_by_node}{An integer indicating the inner node after which the item paramter should be colored}

\item{ABC_colors}{A character vector of length three indicating the colors in which items classified as A/B/C should be displayed}

\item{ABC_size}{A vector of length three indicating the size in which items classified as A/B/C should be displayed}

\item{node_background}{A character vector of length two indicating the background colors of the panels for which the items have been compared}

\item{...}{arguments passed to the underlying functions, i.e., to mob_control for raschtree, and to the underlying predict and plot methods, respectively.}
}
\value{
An object of S3 class "raschtree" inheriting from class "modelparty".
}
\description{
Additional options to show the ETS-Mantel-Haenszel effect size classification in the inner nodes,
option to color the items in terminal nodes by A/B/C DIF classification based on inner nodes,
and option to turn on a background color so you can see which are the left and right terminal nodes for the inner nodes which items are colored
}
