% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mantelhaenszel.R
\name{purify_MH}
\alias{purify_MH}
\title{Performs purification of the Mantel-Haenszel odds ratio in the Delta-scale}
\usage{
purify_MH(dat, group, sums, MH_stat, which_MH = which_MH)
}
\arguments{
\item{dat}{A dataframe with dichotomous item responses}

\item{group}{An indicator of the two groups for which DIF analyses should be performed}

\item{sums}{A sum score used as a matching criterion for computing the Mantel Haenszel statistic}

\item{MH_stat}{A vector or dataframe including the Mantel-Haenszel odds ratio in the Delta-scale}

\item{which_MH}{a vector containing the item numbers of items with Delta-MH > 1}
}
\value{
A list with purified the Mantel-Haenszel odds ratio in the Delta-scale, standard deviation of purified the Mantel-Haenszel odds ratio in the Delta-scale, a vector     indicating with items were purified
}
\description{
Performs purification of the Mantel-Haenszel odds ratio in the Delta-scale
}
