% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_splitGroups.R
\name{return_splitGroups}
\alias{return_splitGroups}
\title{Returns the splitGroups of a raschtree based on the node structure and the allocation of respondents in terminal nodes}
\usage{
return_splitGroups(node, datFitted)
}
\description{
Returns the splitGroups of a raschtree based on the node structure and the allocation of respondents in terminal nodes
}
