% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ETSMH.R
\name{show_ETSMH}
\alias{show_ETSMH}
\title{Helper function to display the ETS Mantel-Haenszel classification summary in each inner node of the Raschtree}
\usage{
show_ETSMH(
  object,
  id = TRUE,
  pval = TRUE,
  abbreviate = FALSE,
  fill = "white",
  gp = gpar()
)
}
\arguments{
\item{object}{An object of class raschtree that has the mantelHaenszel statistic added to it. See @examples}

\item{id}{Argument copied from partykit:::node_inner}

\item{pval}{Argument copied from partykit:::node_inner}

\item{abbreviate}{Argument copied from partykit:::node_inner}

\item{fill}{Argument copied from partykit:::node_inner}

\item{gp}{Argument copied from partykit:::node_inner}
}
\value{
A function that can be used as a value for the argument 'inner_panel' in plot.raschtree()
}
\description{
This function is largely copied from partykit:::node_inner and adapted to display larger inner nodes with corresponding Mantel-Haenszel classification summaries
}
