% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopFun_mantelhaenszel.R
\name{stopfun_mantelhaenszel}
\alias{stopfun_mantelhaenszel}
\title{Returns a function that can be used for the stopfun argument in the Raschtree.
The stopping function returns a logical indicating whether all items are classified as the stopping criterion in the tree growing process
It uses the calculate_mantelhaenszel function for the Mantel-Haenszel effect size measure and classification}
\usage{
stopfun_mantelhaenszel(purification, stopcrit = c("A"))
}
\arguments{
\item{purification}{A character indicating the type of purification should be used? Options "none", "2step", "iterative"}

\item{stopcrit}{A character indicating when the Raschtree should be stopped. Default is "A", hence the Raschtree is stopped when all items fall into category "A" of the ETS classification scheme. When c("A", "B"), the Raschtree is stopped when all items fall in either category "A" or "B".}
}
\value{
A function that can be used for the stopfun argument in the Raschtree
}
\description{
Returns a function that can be used for the stopfun argument in the Raschtree.
The stopping function returns a logical indicating whether all items are classified as the stopping criterion in the tree growing process
It uses the calculate_mantelhaenszel function for the Mantel-Haenszel effect size measure and classification
}
