% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mantelhaenszel.R
\name{summary_mantelhaenszel}
\alias{summary_mantelhaenszel}
\title{Improves the output of the Mantel Haenszel statistic for the raschtree}
\usage{
summary_mantelhaenszel(x)
}
\description{
Improves the output of the Mantel Haenszel statistic for the raschtree
}
